+++ src/main/java/com/sl/service/impl/BookTypeServiceImpl.java
@@ -2,8 +2,12 @@ package com.sl.service.impl;
 
 import com.sl.dao.BookTypeDao;
 import com.sl.entity.BookType;
+import com.sl.exception.NoFindBookType;
+import com.sl.exception.NotNameException;
+import com.sl.exception.NullIdException;
 import com.sl.service.BookTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,22 +26,32 @@ public class BookTypeServiceImpl implements BookTypeService {
 
     @Override
     public List<BookType> getAll() {
-        return  bookTypeDao.getAll();
+        return bookTypeDao.getAll();
     }
 
 
     @Override
-    public void saveType(BookType bookType){
+    public void saveType(BookType bookType) {
         bookTypeDao.saveType(bookType);
     }
 
     @Override
     public void removeType(Integer id) {
+        if (id == null) {
+            throw new NullIdException("输入有误！", HttpStatus.NOT_ACCEPTABLE);
+        }
+        BookType one = bookTypeDao.findOne(id);
+        if (one == null) {
+            throw new NoFindBookType("要删除的类型不存在",HttpStatus.NOT_FOUND);
+        }
         bookTypeDao.removeType(id);
     }
 
     @Override
     public void modifyType(BookType bookType) {
-        bookTypeDao.modifyType(bookType);
+        if (bookType.getName()!=null){
+            bookTypeDao.modifyType(bookType);
+        }
+        throw new NotNameException("修改类型名为空",HttpStatus.NOT_ACCEPTABLE);
     }
 }
